Send legacy default SMS change broadcast in SmsApplication.
We are moving RoleManagerService into mainline, so we can't reference
SmsApplication there without making it an API, and we can't simply
send the broadcast in RoleManagerService because the logic isn't
simple. Instead, we can simply register a role change listener in
SmsApplication for sending the broadcast, similar to how it is already
listening to package changes.

Bug: 158736025
Test: presubmit
Change-Id: I10e440499003f0ec34142a56e715caf452dbfa50